requirements.txt
Brotli==1.1.0
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==3.3.2
cloudscraper==1.2.71
colorama==0.4.6
curl-cffi==0.5.10
idna==3.6
loguru==0.7.2
Naked==0.1.32
pycparser==2.21
pycryptodome==3.20.0
pyparsing==3.1.1
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
shellescape==3.8.1
urllib3==2.0.7
win32-setctime==1.1.0
